+++ tests/test_metrics/conftest.py
@@ -5,14 +5,14 @@ import pytest
 def root_mean_squared_error():
     from bayesflow.metrics import RootMeanSquaredError
 
-    return RootMeanSquaredError(normalize=True, name="rmse", dtype="float64")
+    return RootMeanSquaredError(normalize=True, name="rmse", dtype="float32")
 
 
 @pytest.fixture()
 def maximum_mean_discrepancy():
     from bayesflow.metrics import MaximumMeanDiscrepancy
 
-    return MaximumMeanDiscrepancy(name="mmd", kernel="gaussian", unbiased=True, dtype="float64")
+    return MaximumMeanDiscrepancy(name="mmd", kernel="gaussian", unbiased=True, dtype="float32")
 
 
 @pytest.fixture(params=["root_mean_squared_error", "maximum_mean_discrepancy"])
